task0

+++ 0x0E-structures_typedef/dog.h
@@ -12,12 +12,12 @@
 struct dog
 {
 	char *name;
-	int *age;
+	float *age;
 	char *owner;
 };
 
 typedef struct dog dog_t;
-void init_dog(struct dog *d, int *name, float age, char *owner);
+void init_dog(struct dog *d, float *name, float age, char *owner);
 void print_dog(struct dog *d);
 dog_t *new_dog(char *name, float age, char *owner);
 void free_dog(dog_t *d);
